move code to ensure new translation is only run if opted into, and to reuse virtualservice translation
