+++ projects/gateway/pkg/translator/http_translator_test.go
@@ -26,6 +26,9 @@ import (
 
 var _ = Describe("Http Translator", func() {
 
+	// This file contains the tests both for the HttpTranslator and VirtualServiceTranslator
+	// It would be ideal to split the VirtualServiceTranslator tests into a distinct file
+
 	var (
 		ctx        context.Context
 		cancel     context.CancelFunc
@@ -39,7 +42,11 @@ var _ = Describe("Http Translator", func() {
 	BeforeEach(func() {
 		ctx, cancel = context.WithCancel(context.Background())
 
-		translator = &HttpTranslator{}
+		translator = &HttpTranslator{
+			VirtualServiceTranslator: &VirtualServiceTranslator{
+				WarnOnRouteShortCircuiting: false,
+			},
+		}
 	})
 
 	JustBeforeEach(func() {
@@ -547,7 +554,9 @@ var _ = Describe("Http Translator", func() {
 
 			BeforeEach(func() {
 				translator = &HttpTranslator{
-					WarnOnRouteShortCircuiting: true,
+					VirtualServiceTranslator: &VirtualServiceTranslator{
+						WarnOnRouteShortCircuiting: true,
+					},
 				}
 			})
