+++ projects/gateway/pkg/translator/listener_translator.go
@@ -15,7 +15,6 @@ var MissingGatewayTypeErr = errors.New("invalid gateway: gateway must contain ga
 
 // ListenerTranslator converts a Gateway into a Listener
 type ListenerTranslator interface {
-	Name() string
 	ComputeListener(params Params, proxyName string, gateway *v1.Gateway) *gloov1.Listener
 }
 
@@ -34,10 +33,6 @@ func NewTranslatorParams(ctx context.Context, snapshot *v1.ApiSnapshot, reports 
 }
 
 type InvalidGatewayTypeTranslator struct{}
-
-func (n InvalidGatewayTypeTranslator) Name() string {
-	return "invalid-gateway-type"
-}
 
 func (n InvalidGatewayTypeTranslator) ComputeListener(params Params, proxyName string, gateway *v1.Gateway) *gloov1.Listener {
 	params.reports.AddError(gateway, MissingGatewayTypeErr)
