+++ projects/gateway/pkg/translator/opts.go
@@ -7,23 +7,23 @@ import (
 )
 
 type Opts struct {
-	GlooNamespace                 string
-	WriteNamespace                string
-	StatusReporterNamespace       string
-	WatchNamespaces               []string
-	Gateways                      factory.ResourceClientFactory
-	MatchableHttpGateways         factory.ResourceClientFactory
-	VirtualServices               factory.ResourceClientFactory
-	RouteTables                   factory.ResourceClientFactory
-	Proxies                       factory.ResourceClientFactory
-	RouteOptions                  factory.ResourceClientFactory
-	VirtualHostOptions            factory.ResourceClientFactory
-	WatchOpts                     clients.WatchOpts
-	ValidationServerAddress       string
-	DevMode                       bool
-	ReadGatewaysFromAllNamespaces bool
-	Validation                    *ValidationOpts
-	ConfigStatusMetricOpts        map[string]*metrics.Labels
+	GlooNamespace                  string
+	WriteNamespace                 string
+	StatusReporterNamespace        string
+	WatchNamespaces                []string
+	Gateways                       factory.ResourceClientFactory
+	MatchableHttpGateways          factory.ResourceClientFactory
+	VirtualServices                factory.ResourceClientFactory
+	RouteTables                    factory.ResourceClientFactory
+	Proxies                        factory.ResourceClientFactory
+	RouteOptions                   factory.ResourceClientFactory
+	VirtualHostOptions             factory.ResourceClientFactory
+	WatchOpts                      clients.WatchOpts
+	ValidationServerAddress        string
+	DevMode                        bool
+	ReadGatewaysFromAllNamespaces  bool
+	Validation                     *ValidationOpts
+	ConfigStatusMetricOpts         map[string]*metrics.Labels
 	IsolateVirtualHostsBySslConfig bool
 }
